fix: prefix url attributes for current page (#47)
Potentially breaking change:
Prefixes the url attributes describing the current page with
"page." in order to deconflict their usage with events emitted
by the fetch instrumentation. This enables the fetch instrumentation
to now also add url attributes for the current page not just
the target url.

This should not require any adjustments on the implementation side,
but might require updates to existing dashboards or saved views