feat: Add xray propagator and introduce propagator config (#54)
The SDK now supports AWS context propagation by injecting the `X-Amzn-Trace-Id` header if configured. 
This change comes with a deprecation of the old `propagateTraceHeadersCorsURLs` config and a new propagator config option. 
Each propagator can now be configured separately for cross-origin requests. Same origin request always add the `traceparent` and if configured the `X-Amzn-Trace-Id` headers. 
Example config: 
```js
init({
  propagators: [
    { type: "traceparent", match: [/.*\/api\/special.*/] },
    { type: "xray", match: [/.*\/api\/special.*/] },
  ],
});
```
We recommend to migrate the existing config to the new schema since the old config will be removed in future versions.