fix(fetch): do not wrap body for responses that cant have a body
Responses with some status codes like 204 and 304 can't have a
response body. Browsers are still providing a readable stream for
them when returning the response from fetch, but a new response
can't be constructed with a body from js for these status codes.
This lead to some exceptions when instrumenting requests with
responses carrying those status codes.

This adds a short cuircit to the response wrapping logic to avoid
responses where we could not construct a copy with a wrapped body.
Generally these should not have a body anyways so we don't need to
try waiting for the body to be read either.